prov/rxm: Updates to rxm provider to suppress warnings

+++ prov/rxm/src/rxm.h
@@ -290,7 +290,7 @@ struct rxm_recv_queue {
 struct rxm_buf_pool {
 	struct util_buf_pool *pool;
 	struct dlist_entry buf_list;
-	uint8_t local_mr;
+	int local_mr;
 	fastlock_t lock;
 };
 
@@ -355,7 +355,7 @@ int rxm_domain_open(struct fid_fabric *fabric, struct fi_info *info,
 int rxm_cq_open(struct fid_domain *domain, struct fi_cq_attr *attr,
 			 struct fid_cq **cq_fid, void *context);
 void rxm_cq_progress(struct rxm_ep *rxm_ep);
-int rxm_cq_handle_data(struct rxm_rx_buf *rx_buf);
+ssize_t rxm_cq_handle_data(struct rxm_rx_buf *rx_buf);
 
 int rxm_endpoint(struct fid_domain *domain, struct fi_info *info,
 			  struct fid_ep **ep, void *context);
